requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
html-testRunner==1.2.1
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.7
parameterized==0.7.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
toml==0.10.2
urllib3==1.26.2
webdriver-manager==3.2.2
